requirements.txt
Django==5.2.6
djangorestframework==3.16.1
Pillow==11.3.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2

# Production/deployment
gunicorn==21.2.0
whitenoise==6.7.0
dj-database-url==2.3.0
psycopg2-binary==2.9.9
